Fat Steve has a voice line! + LOTS of decoration
Fat Steve plays his voice line with "professionally voiced dialogue" when the player enters his trigger zone (which is huge so odds are you'll only hear it once). Also added tons of decoration and life to the scene. There are now 3 floor types: street, interior, and roof(unchanged, it looked good as is).